РџСЂРѕРІРµСЂРєР° РѕР±С‰РµР№ РєРѕРЅС‚СЂРѕР»СЊРЅРѕР№ СЃСѓРјРјС‹. РћС‚РІРµС‚ РЅР° Р±СЂРѕР°РґРєР°СЃС‚ СЃ РїСЂР°РІРёР»СЊРЅС‹Рј РёРґРµРЅС‚РёС„РёРєР°С‚РѕСЂРѕРј.

+++ Fudp/FudpMessage/proglist.cpp
@@ -26,7 +26,7 @@ std::vector<byte> ProgList::encode()
         for (int i =0; i < t.size(); i ++)
             in << (quint8)t.at(i);
 
-        in << i->getFileSize() << i->getControlSum();
+        in << i->getFileSize() << (quint32) i->getControlSum();
     }
     return Message::fromQByteArrayToVector(buffer);
 }
